+++ litellm/utils.py
@@ -7661,7 +7661,7 @@ class ProviderConfigManager:
 
             return VertexAIVideoConfig()
         elif LlmProviders.RUNWAYML == provider:
-            from litellm.llms.runway.videos.transformation import RunwayMLVideoConfig
+            from litellm.llms.runwayml.videos.transformation import RunwayMLVideoConfig
 
             return RunwayMLVideoConfig()
         return None
